Решение задачи "Java HashMap: Поисковый мини-движок Google". Исправления: вынес работу с веб-страницами в отдельный класс, добавил проверку на null, добавил аннотацию data в WebPage.
