+++ src/main/java/searchengine/Main.java
@@ -1,16 +1,11 @@
 package searchengine;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
+import static searchengine.WebPageService.getWebPagesByKeyword;
+import static searchengine.WebPageService.indexTheWebPage;
+import static searchengine.WebPageService.removeWebPageByUrl;
 
 public class Main {
-    private static Map<String, List<WebPage>> index = new HashMap<>();
-    private static Map<String, List<String>> urlAndKeywords = new HashMap<>();
+
 
     public static void main(String[] args) {
         indexTheWebPage(new WebPage("url1", "title1", "hello, world!"));
@@ -28,45 +23,5 @@ public class Main {
         indexTheWebPage(new WebPage("url1", "title1", "hello, world!"));
         System.out.println(getWebPagesByKeyword("hello"));
         System.out.println(getWebPagesByKeyword("world"));
-    }
-
-    public static void indexTheWebPage(WebPage webPage) {
-        if (urlAndKeywords.containsKey(webPage.getUrl())) {
-            return;
-        }
-        List<String> keywordsByNewUrl = new ArrayList<>();
-
-        String[] words = webPage.getContent().split("\\P{L}+");
-        Set<String> uniqueKeywords = new HashSet<>(Arrays.asList(words));
-        for (var keyword : uniqueKeywords) {
-            if (index.containsKey(keyword)) {
-                index.get(keyword).add(webPage);
-            } else {
-                List<WebPage> newKeywordWebPages = new ArrayList<>();
-                newKeywordWebPages.add(webPage);
-                index.put(keyword, newKeywordWebPages);
-            }
-            keywordsByNewUrl.add(keyword);
-        }
-        urlAndKeywords.put(webPage.getUrl(), keywordsByNewUrl);
-    }
-
-    public static List<WebPage> getWebPagesByKeyword(String keyword) {
-        return index.get(keyword);
-    }
-
-    public static void removeWebPageByUrl(String url) {
-        for (var keyword : urlAndKeywords.get(url)) {
-            for (int i = 0; i < index.get(keyword).size(); i++) {
-                if (index.get(keyword).get(i).getUrl().equals(url)) {
-                    index.get(keyword).remove(i);
-                    if (index.get(keyword).isEmpty()) {
-                        index.remove(keyword);
-                    }
-                    break;
-                }
-            }
-        }
-        urlAndKeywords.remove(url);
     }
 }
