+++ src/main/java/searchengine/WebPage.java
@@ -1,15 +1,13 @@
 package searchengine;
 
 import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.ToString;
+import lombok.Data;
 
+@Data
 @AllArgsConstructor
-@ToString
 public class WebPage {
-    @Getter
     private String url;
     private String title;
-    @Getter
     private String content;
 }
+
